src/hub/requirements.txt
python-dotenv
numpy
librosa
flask
pydub
